fix: address code review findings
- Fix chargingChanged always-false latent bug (compute before updating wasCharging)
- Remove NVS write from CannedMessages::next() to match original behavior (save on sleep/poweroff only)
- Fix narrowing cast in port comparison (uint8_t -> uint32_t to match port type)

+++ src/protocol/MeshPacket.cpp
@@ -63,7 +63,7 @@ MeshError parsePacket(const uint8_t* frame, size_t frameLen,
         return MeshError::DecodeFailed;
     }
 
-    if (port != static_cast<uint8_t>(PortNum::TextMessage) || !payload || payloadLen == 0) {
+    if (port != static_cast<uint32_t>(PortNum::TextMessage) || !payload || payloadLen == 0) {
         // Not a text message — not an error, just nothing to extract
         return MeshError::Ok;
     }
